backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.3.0",
    "flask-cors>=4.0.0",
    "requests>=2.32.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
    "fish-audio-sdk>=1.0.0",
    "numpy>=1.26.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
    "python-socketio[client]>=5.11.0",
]
